planner refactoring of operators
Changes how we push predicates down the operator tree. We now have a
single predicate pushing method for all operators.
