+++ go/vt/vtgate/planbuilder/operators/union.go
@@ -17,7 +17,10 @@ limitations under the License.
 package operators
 
 import (
+	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/sqlparser"
+	"vitess.io/vitess/go/vt/vterrors"
+	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
 )
 
 type Union struct {
@@ -45,4 +48,53 @@ func (u *Union) Clone(inputs []Operator) Operator {
 // Inputs implements the Operator interface
 func (u *Union) Inputs() []Operator {
 	return u.Sources
+}
+
+func (u *Union) addPredicate(ctx *plancontext.PlanningContext, expr sqlparser.Expr) error {
+	offsets := make(map[string]int)
+	for i, selectExpr := range u.SelectStmts[0].SelectExprs {
+		ae, ok := selectExpr.(*sqlparser.AliasedExpr)
+		if !ok {
+			return vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, "can't push predicates on UNION where the first SELECT contains star or next")
+		}
+		if !ae.As.IsEmpty() {
+			offsets[ae.As.String()] = i
+			continue
+		}
+		col, ok := ae.Expr.(*sqlparser.ColName)
+		if ok {
+			offsets[col.Name.Lowered()] = i
+		}
+	}
+
+	for i := range u.Sources {
+		predicate := sqlparser.CloneExpr(expr)
+		var err error
+		predicate = sqlparser.Rewrite(predicate, func(cursor *sqlparser.Cursor) bool {
+			col, ok := cursor.Node().(*sqlparser.ColName)
+			if !ok {
+				return err == nil
+			}
+
+			idx, ok := offsets[col.Name.Lowered()]
+			if !ok {
+				err = vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, "can't push predicates on concatenate")
+				return false
+			}
+
+			ae, ok := u.SelectStmts[i].SelectExprs[idx].(*sqlparser.AliasedExpr)
+			if !ok {
+				err = vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, "can't push predicates on concatenate")
+				return false
+			}
+			cursor.Replace(ae.Expr)
+			return false
+		}, nil).(sqlparser.Expr)
+		if err != nil {
+			return err
+		}
+		u.Sources[i], err = PushPredicate(ctx, predicate, u.Sources[i])
+	}
+
+	return nil
 }
