+++ go/vt/vtgate/planbuilder/operators/vindex.go
@@ -21,6 +21,7 @@ import (
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vterrors"
 	"vitess.io/vitess/go/vt/vtgate/engine"
+	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
 	"vitess.io/vitess/go/vt/vtgate/semantics"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
 )
@@ -88,5 +89,48 @@ func (v *Vindex) CheckValid() error {
 		return vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, "unsupported: where clause for vindex function must be of the form id = <val> or id in(<val>,...) (where clause missing)")
 	}
 
+	return nil
+}
+
+func (v *Vindex) addPredicate(ctx *plancontext.PlanningContext, expr sqlparser.Expr) error {
+	for _, e := range sqlparser.SplitAndExpression(nil, expr) {
+		deps := ctx.SemTable.RecursiveDeps(e)
+		if deps.NumberOfTables() > 1 {
+			return vterrors.Errorf(vtrpcpb.Code_UNIMPLEMENTED, vindexUnsupported+" (multiple tables involved)")
+		}
+		// check if we already have a predicate
+		if v.OpCode != engine.VindexNone {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (multiple filters)")
+		}
+
+		// check LHS
+		comparison, ok := e.(*sqlparser.ComparisonExpr)
+		if !ok {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (not a comparison)")
+		}
+		if comparison.Operator != sqlparser.EqualOp && comparison.Operator != sqlparser.InOp {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (not equality)")
+		}
+		colname, ok := comparison.Left.(*sqlparser.ColName)
+		if !ok {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (lhs is not a column)")
+		}
+		if !colname.Name.EqualString("id") {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (lhs is not id)")
+		}
+
+		// check RHS
+		var err error
+		if sqlparser.IsValue(comparison.Right) || sqlparser.IsSimpleTuple(comparison.Right) {
+			v.Value = comparison.Right
+		} else {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+" (rhs is not a value)")
+		}
+		if err != nil {
+			return vterrors.Errorf(vtrpcpb.Code_INTERNAL, vindexUnsupported+": %v", err)
+		}
+		v.OpCode = engine.VindexMap
+		v.Table.Predicates = append(v.Table.Predicates, e)
+	}
 	return nil
 }
